[Driver] Add flag to force file lists to not be used
When generating a compiler invocation via driver::createCompilerInvocation()
the we end up using file lists if the number of inputs is > 128. We don't
actually write out the file list along this path though, so callers (SourceKit)
will fail when we try to generate an AST.

This patch adds a flag to force not using file lists irrespective of whether
-driver-use-filelists is specified or the number of inputs is >128 (none of
its callers care about argv limits).

Resolves rdar://problem/38231888